Add the most basic form of Rate Limiting. Saves data to a JSON file.
Make attaching middlewares to routes with instantiating a new object of the middleware instead of just passing in the class name.

default JSON class to return an empty array if we can't decode the string.

Add a new RateLimited Response.

Allow killing the request early.

Add File DataSource to allow easier working with files.

A few more cool touches.